Move Logic from component to a service
