+++ feature-libs/pickup-in-store/components/container/pickup-in-store-details/pickup-in-store-details.component.html
@@ -12,4 +12,4 @@ Pick In Store Details
   </div>
 </ng-container>
 
-<!-- <pre>{{ deliveryPointsOfService$ | async | json }}</pre> -->
+
