+++ feature-libs/pickup-in-store/root/model/delivery-points.model.ts
@@ -1,0 +1,11 @@
+
+import { OrderEntry } from '@spartacus/cart/base/root';
+import { PointOfService } from '@spartacus/core';
+
+export type DeliveryPointOfServiceItems = {
+    name: string;
+    storeDetails: PointOfService;
+    value: Array<OrderEntry>;
+}
+
+
